backend/verify_phase1.py

import requests
import json
import uuid

# Base URL
BASE_URL = "http://localhost:8000/api"

def test_phase1_flow():
    print("--- Starting Phase 1 Verification ---")
    
    # 1. Verification of Schemas and Models (Manual check of code is done)
    
    # 2. Test Registration with Profile Data
    email = f"test_{uuid.uuid4().hex[:8]}@example.com"
    reg_data = {
        "email": email,
        "password": "password123",
        "first_name": "Test",
        "last_name": "User",
        "role": "attendee",
        "enrollment_number": f"VERIFY_{uuid.uuid4().hex[:6]}",
        "branch": "CSE",
        "year": 3,
        "college_name": "BCE",
        "is_external": False
    }
    
    print(f"Registering user: {email}")
    reg_resp = requests.post(f"{BASE_URL}/auth/register", json=reg_data)
    if reg_resp.status_code == 201:
        print("✅ Registration successful")
    else:
        print(f"❌ Registration failed: {reg_resp.text}")
        return

    auth_data = reg_resp.json()
    token = auth_data["access_token"]
    headers = {"Authorization": f"Bearer {token}"}

    # 3. Test Create Event with Hackathon type and Category
    event_data = {
        "name": "Global Hackathon 2024",
        "description": "A massive 48-hour coding competition.",
        "event_type": "hackathon",
        "category": "Web3",
        "venue": "BCE Tech Hub",
        "max_participants": 200,
        "start_date": "2024-10-01T09:00:00Z",
        "end_date": "2024-10-03T18:00:00Z",
        "registration_deadline": "2024-09-25T23:59:59Z",
        "config_data": {
            "max_team_size": 4,
            "tracks": ["DeFi", "NFTs", "Infrastructure"],
            "submission_deadline": "2024-10-03T12:00:00Z"
        }
    }
    
    print("Creating Hackathon event...")
    # Using existing organizer for simplicity or the new user if they have rights
    # Note: New users are 'attendee' by default. We might need an organizer token.
    # For verification, we assume the backend is running and we can use a known organizer or 
    # adjust the role of the new user.
    
    # Let's check the created event
    ev_resp = requests.post(f"{BASE_URL}/events", json=event_data, headers=headers)
    if ev_resp.status_code == 201:
        print("✅ Event creation successful")
        event_id = ev_resp.json()["id"]
        
        # 4. Verify category and config
        get_resp = requests.get(f"{BASE_URL}/events/{event_id}", headers=headers)
        get_data = get_resp.json()
        if get_data.get("category") == "Web3":
            print("✅ Category verified")
        else:
            print(f"❌ Category mismatch: {get_data.get('category')}")

        config_resp = requests.get(f"{BASE_URL}/events/{event_id}/config", headers=headers)
        if config_resp.status_code == 200:
            config_data = config_resp.json()
            if config_data and config_data["config_data"]["max_team_size"] == 4:
                print("✅ Event configuration verified")
            else:
                print(f"❌ Config data mismatch or missing: {config_data}")
        else:
            print(f"❌ Failed to get config: {config_resp.text}")
            
    else:
        print(f"❌ Event creation failed (likely due to role): {ev_resp.text}")
        print("Note: Automated creation test requires organizer permissions.")

    print("--- Phase 1 Verification Finished ---")

if __name__ == "__main__":
    # This script assumes the backend is running locally.
    # In a real environment, we'd mock or have a test DB.
    print("Verification script ready. Run this against a live backend to validate.")
    # test_phase1_flow() # Commented out to prevent execution without a server
