backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Rate Limiting
slowapi==0.1.9

# Supabase
supabase==2.11.0

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# Security
python-jose[cryptography]==3.3.0

# Data & Utilities
python-dotenv==1.0.0
httpx==0.27.0
email-validator==2.1.0.post1

# PDF Generation
reportlab==4.0.9
jinja2==3.1.3

# Date handling
python-dateutil==2.8.2

# WebSocket
python-socketio==5.11.0
bidict==0.23.1
